handle non-struct types properly
this is a bit of a hack. WriteTo is always defined on a pointer
receiver for efficiency reasons, but this also implies that it
should be valid for nil pointers. We could use generatePointer
and preface every encoding with either 0 or 1, but that strikes
me as ugly. Need to think about this some more.